requirements.txt
# Core dependencies
openai>=1.10.0        # Optional: Only needed when using the 'openai' provider
requests>=2.31.0      # For HTTP requests
beautifulsoup4>=4.12.2 # For HTML parsing
python-dotenv>=1.0.0  # For environment variables
tenacity>=8.2.3       # For API error retries
markdownify>=0.11.6   # For HTML to Markdown conversion
langchain>=0.1.0      # For LLM chains
colorlog>=6.7.0       # For colored logging
aiohttp>=3.9.0        # For async HTTP requests
httpx>=0.25.0         # For async HTTP requests (Wikipedia tool)
aiosqlite>=0.19.0     # For SQLite cache

# For data processing and analysis
numpy>=1.24.0         # For numerical calculations
pandas>=2.0.0         # For tabular data processing
tqdm>=4.65.0          # For progress display
scikit-learn>=1.3.0   # For cosine similarity

# For enhanced search and bypassing blocks
selenium>=4.10.0      # For browser automation
fake-useragent>=1.1.3 # For User-Agent rotation
requests-cache>=1.0.0 # For HTTP request caching
fake-headers>=1.0.2   # For generating realistic headers

# For text and language processing
spacy>=3.5.0          # For NLP processing
langdetect>=1.0.9     # For text language detection
sentence-transformers>=2.2.2 # For semantic search embeddings

# Specialized Search Tools
arxiv>=2.1.0           # For ArXiv search
youtube-search-python>=1.6.6 # For YouTube search

# Web Interface
fastapi>=0.104.0       # For web API
uvicorn>=0.24.0        # For ASGI server
websockets>=12.0       # For WebSocket support

# Optional provider-specific dependencies
# anthropic>=0.7.0     # Only needed when using the 'anthropic' provider
ollama>=0.1.6          # For Ollama integration
